feat: Move tests to the items_request_spec instead of item_spec
