+++ spec/models/item_spec.rb
@@ -24,55 +24,55 @@ RSpec.describe Item, type: :model do
     end
   end
 
-  describe ' find all items based on min price #find_items_by_price' do
-    xit 'returns all items that match a minimum price parameter' do
-      get '/api/v1/items/find_all', params: { min_price: 50.00 }
+  # describe ' find all items based on min price #find_items_by_price' do
+  #   it 'returns all items that match a minimum price parameter' do
+  #     get "/api/v1/items/find_all", params: { min_price: 50.00 }
   
-      expect(response).to be_successful
+  #     expect(response).to be_successful
     
-      itmes = JSON.parse(response.body, symbolize_names: true)[:data]
-      expect(items.count).to eq(2)
+  #     itmes = JSON.parse(response.body, symbolize_names: true)[:data]
+  #     expect(items.count).to eq(2)
     
-      expect(items).to eq([@item1, @item3])
-    end
-  end
+  #     expect(items).to eq([@item1, @item3])
+  #   end
+  # end
 
-  describe ' find all items based on max price #find_items_by_price' do
-    xit 'returns all items that match a maximum price parameter' do
-      get '/api/v1/items/find_all', params: { max_price: 70.00 }
+  # describe ' find all items based on max price #find_items_by_price' do
+  #   it 'returns all items that match a maximum price parameter' do
+  #     get '/api/v1/items/find_all', params: { max_price: 70.00 }
   
-      expect(response).to be_successful
+  #     expect(response).to be_successful
     
-      itmes = JSON.parse(response.body, symbolize_names: true)[:data]
-      expect(items.count).to eq(2)
+  #     itmes = JSON.parse(response.body, symbolize_names: true)[:data]
+  #     expect(items.count).to eq(2)
     
-      expect(items).to include(
-        include(
-          attributes: include(name: "hydroflask", unit_price: 20.00)
-        ),
-        include(
-          attributes: include(name: "airpump", unit_price: 55.00)
-        )
-      )  
-    end
-  end
+  #     expect(items).to include(
+  #       include(
+  #         attributes: include(name: "hydroflask", unit_price: 20.00)
+  #       ),
+  #       include(
+  #         attributes: include(name: "airpump", unit_price: 55.00)
+  #       )
+  #     )  
+  #   end
+  # end
 
-  describe ' find all itmes using #find_items_by_name' do
-    xit 'returns all itmes that match a name' do
-      get '/api/v1/items/find_all', params: { name: 'air' }
+  # describe ' find all itmes using #find_items_by_name' do
+  #   it 'returns all itmes that match a name' do
+  #     get '/api/v1/items/find_all', params: { name: 'air' }
 
-      expect(response).to be_successful
+  #     expect(response).to be_successful
   
-      items = JSON.parse(response.body, symbolize_names: true)[:data]
-      expect(items.count).to eq(2) 
-      expect(items).to include(
-        include(
-          attributes: include(name: "airbuds", price: 75.00)
-        ),
-        include(
-          attributes: include(name: "airpump", price: 55.00)
-        )
-      )
-    end
-  end
+  #     items = JSON.parse(response.body, symbolize_names: true)[:data]
+  #     expect(items.count).to eq(2) 
+  #     expect(items).to include(
+  #       include(
+  #         attributes: include(name: "airbuds", price: 75.00)
+  #       ),
+  #       include(
+  #         attributes: include(name: "airpump", price: 55.00)
+  #       )
+  #     )
+  #   end
+  # end
 end
